+++ Changelog.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project follows to [Ragnarök Versioning Convention](https://github.com/Red-Studio-Ragnarok/Commons/blob/main/Ragnar%C3%B6k%20Versioning%20Convention.md).
 
-## [UNRELEASED] Red Core Version 0.6 Changelog
+## Red Core Version 0.6 Changelog - 2025-02-01
 
 ### Deprecation Warning
 
@@ -28,6 +28,7 @@ The method `RedLogger#printFramedError` has been renamed to `RedLogger#logFramed
 - Improved documentation for `RedLogger#logFramedError`
 - Renamed `RedLogger#printFramedError` to `RedLogger#framedError`
 - Switched from Javadoc to Markdowndoc
+- Most utility classes now use Lombok `@UtilityClass`
 
 ### Internal